Upstage Studio V1.1 (#681)
* migrate from app1 to upstage.live (#671)

* hotfix blank create stage screen (#674)

* authorization in graphql request

* authorization and refresh token logic in apollo

* fix notification and messages not appearing

* multiframe preview and delete media

* media position swapable using drag gesture

* media form layout, remove multiframe

* upload file and create new media

* edit media interface

* fix build failed at tsc

* timezone

+++ stage/asset.py
@@ -320,7 +320,7 @@ class UpdateMedia(graphene.Mutation):
 
 
 class DeleteMedia(graphene.Mutation):
-    """Mutation to sweep a stage."""
+    """Mutation to delete a media."""
     success = graphene.Boolean()
     message = graphene.String()
 
@@ -342,15 +342,16 @@ class DeleteMedia(graphene.Mutation):
 
                 if asset.description:
                     attributes = json.loads(asset.description)
-                    # Delete frames that was assigned only to this media
-                    for frame in attributes['frames']:
-                        frame_asset = local_db_session.query(AssetModel).filter(
-                            AssetModel.file_location == frame).first()
-                        if not frame_asset:
-                            physical_path = os.path.join(
-                                absolutePath, storagePath, frame)
-                            if os.path.exists(physical_path):
-                                os.remove(physical_path)
+                    if 'frames' in attributes and attributes['frames']:
+                        # Delete frames that was assigned only to this media
+                        for frame in attributes['frames']:
+                            frame_asset = local_db_session.query(AssetModel).filter(
+                                AssetModel.file_location == frame).first()
+                            if not frame_asset:
+                                physical_path = os.path.join(
+                                    absolutePath, storagePath, frame)
+                                if os.path.exists(physical_path):
+                                    os.remove(physical_path)
 
                 physical_path = os.path.join(
                     absolutePath, storagePath, asset.file_location)
